Add DataSource module and re-export
